Repaired permissions

+++ del__car.php
@@ -1,7 +1,14 @@
 <?php
+    session_start();
     include_once('connect.php');
-    $car_id = $_POST['car__id'];
-    $sql = "DELETE FROM cars WHERE id = $car_id";
-    mysqli_query($mysqli,$sql);
-    header('location:login.php');
+    include_once('functions.php');
+
+    if(checkPrivileges($mysqli,3) == true) {
+        $car_id = $_POST['car__id'];
+        $sql = "DELETE FROM cars WHERE id = $car_id";
+        mysqli_query($mysqli,$sql);
+        header('location:login.php');
+    } else {
+        header('location:login.php?action=no_permission');
+    }
 ?>
